refactor: move level-up.ts
